add presentation layer app-module
