--- app/src/main/java/com/example/compmoviedb/MainActivity.kt
@@ -1,29 +1,0 @@
-package com.example.compmoviedb
-
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material.MaterialTheme
-import androidx.compose.material.Surface
-import androidx.compose.material.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.navigation.compose.rememberNavController
-import com.example.compmoviedb.navigation.SetupNavHost
-import com.example.compmoviedb.screens.MainViewModel
-import com.example.compmoviedb.ui.theme.CompmoviedbTheme
-
-class MainActivity : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContent {
-            CompmoviedbTheme {
-                val navController = rememberNavController()
-                SetupNavHost(navController = navController)
-            }
-        }
-    }
-}
