+++ app/src/main/java/com/example/compmoviedb/presentation/screens/MainScreen.kt
@@ -1,4 +1,4 @@
-package com.example.compmoviedb.screens
+package com.example.compmoviedb.presentation.screens
 
 import androidx.compose.runtime.Composable
 import androidx.navigation.NavController
