+++ app/src/main/java/com/example/compmoviedb/presentation/ui/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.example.compmoviedb.ui.theme
+package com.example.compmoviedb.presentation.ui.theme
 
 import androidx.compose.ui.graphics.Color
