+++ app/src/main/java/com/example/compmoviedb/presentation/ui/theme/Shape.kt
@@ -1,4 +1,4 @@
-package com.example.compmoviedb.ui.theme
+package com.example.compmoviedb.presentation.ui.theme
 
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.Shapes
