+++ app/src/main/java/com/example/compmoviedb/presentation/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.example.compmoviedb.ui.theme
+package com.example.compmoviedb.presentation.ui.theme
 
 import androidx.compose.material.Typography
 import androidx.compose.ui.text.TextStyle
